pyproject.toml
[tool.poetry]
name = "cufeed-py-server"
version = "0.1.0"
description = ""
authors = ["freezner <freezner3@gmail.com>"]
readme = "README.md"
packages = [{include = "cufeed_app"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
pymysql = "^1.1.0"
feedparser = "^6.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
